fixed tests

+++ unit-tests/Int16Test.cc
@@ -127,7 +127,7 @@ public:
 
     void val2buf_test()
     {
-        int i = 42;
+        short i = 42;
         i2->val2buf(&i, true);
         CPPUNIT_ASSERT(i2->value() == 42);        
         CPPUNIT_ASSERT_THROW(i2->val2buf(NULL, true), InternalErr);
@@ -135,7 +135,7 @@ public:
 
     void buf2val_test()
     {
-        int i = 42;
+        short i = 42;
         void *v = &i;
         short *v2 = NULL;
         CPPUNIT_ASSERT(i2->set_value(0));
